+++ src/components/ProgressBar/ProgressBar.stories.tsx
@@ -1,5 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react';
 import { ProgressBar } from './ProgressBar';
+import { steps } from './Steps';
 
 const meta = {
     title: 'Components/ProgressBar',
@@ -8,6 +9,9 @@ const meta = {
     parameters: {
         layout: 'padded',
     },
+    argTypes: {
+        activeStep: {},
+    },
 } as Meta<typeof ProgressBar>;
 
 export default meta;
@@ -15,6 +19,20 @@ type Story = StoryObj<typeof meta>;
 
 export const Primary: Story = {
     args: {
-        activeStep: 4,
+        activeStep: 5,
+        steps: steps,
+    },
+    argTypes: {
+        steps: {
+            fields: [
+                'itemTemplate',
+                'wpId',
+                'serialNumber',
+                'vendorId',
+                'isBatch',
+                'preCheck',
+                'documentation',
+            ],
+        },
     },
 };
