+++ src/components/TopBar/TopBar.stories.tsx
@@ -1,11 +1,11 @@
 import { Meta, StoryObj } from '@storybook/react';
-import { TopBar } from './TopBar';
+import { ResponsiveAppBar } from './ResponsiveAppBar';
 
 const meta = {
     title: 'Components/TopBar',
-    component: TopBar,
+    component: ResponsiveAppBar,
     tags: ['autodocs'],
-} as Meta<typeof TopBar>;
+} as Meta<typeof ResponsiveAppBar>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
